requirements.txt
flask
gunicorn
numpy<2
pillow
joblib
scikit-learn
tensorflow-cpu==2.13.0